backend/utils/yaml_handler.py
from __future__ import annotations

from pathlib import Path
from typing import Any

import yaml


def dump_yaml(path: Path, payload: Any) -> None:
    path.parent.mkdir(parents=True, exist_ok=True)
    with path.open("w", encoding="utf-8") as target:
        yaml.safe_dump(payload, target, sort_keys=False)


def load_yaml(path: Path) -> Any | None:
    if not path.exists():
        return None
    with path.open("r", encoding="utf-8") as source:
        return yaml.safe_load(source)